Russian Localization (Not fully finished)
I did about 80% of all the work, made some kind of localization file, shoveled a bunch of text, and so on. The translation is not 100% correct and reliable, I am even sure of it, and I really want someone who understands this more and better than me to do it, I have already" built "a"launch pad" for this.
I worked on all this for 3-4 days and was very tired. I have been continuously translating various mods for about two weeks and I think that it is worth taking a little vacation.
Please check all the lines that I have edited and tell me what is allowed to translate, what is not allowed, and how to do it correctly.